Separate notification permission status from permission actions
