+++ src/screens/SettingsSheet/components/NotificationsSection.tsx
@@ -15,7 +15,8 @@ import useAppState from '@/hooks/useAppState';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
-import { isNotificationPermissionGranted, requestNotificationPermission } from '@/notifications/permissions';
+import { requestNotificationPermission } from '@/notifications/permissions';
+import { isNotificationPermissionGranted } from '@/notifications/permissionStatus';
 import { WalletNotificationRelationship } from '@/notifications/settings/constants';
 import { useAllNotificationSettingsFromStorage, useWalletGroupNotificationSettings } from '@/notifications/settings/hooks';
 import { toggleGlobalNotificationTopic } from '@/notifications/settings/settings';
